ewc: Inline ewc-objects interface functions and make them settable
The interface functions can now be used as gv place forms (setf).
Because gv-get can used the compiler macro. It so so for cl-defstruct accessors.

define-inline lets ewc-objects-id->data and ewc-objects-path->data be used as gv place forms.
Setting path this way is explicitly forbidden and produces an error.
This is done by definin a gv expander for it. Like cl-defstruct does it.

ewc-objects--nth does not have to be inlined for this to work.
I did it anyway for consistency.

cl-defsubst could be used instead of define-inline. It works and has
less boilerplate. But it is apparently broken and not reliable.